+++ oflib/ofl-structs.h
@@ -115,8 +115,8 @@ struct ofl_instruction_meter {
 /* XFSM: for supporting set state that should write the new state in the statetable */
 struct ofl_instruction_set_state {
     struct ofl_instruction_header   header; /* OFPIT_METER */
-    //uint32_t state;
-    uint64_t state;
+    uint32_t state;
+    //uint64_t state;
 };
 
 /* Instruction structure for experimental instructions */
